create file PointController
